# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "学伴 API",
    "description": "学伴（XueBan）AI 学习软件后端服务",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/tutor/session": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "开启守护型讲解会话（F-11）",
        "description": "创建讲解会话（初始层级 0，不直接给答案）。",
        "operationId": "create_session_v1_tutor_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorSessionStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tutor/{session_id}/hint": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "请求下一层提示（不可跳层）",
        "description": "按状态机返回下一层提示（思路 → 关键步骤 → 全解）。",
        "operationId": "request_hint_v1_tutor__session_id__hint_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorHintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TutorHintResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tutor/{session_id}/hint/stream": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "请求下一层提示（SSE 流式）",
        "description": "SSE 流式输出：start → delta* → done（校验在流开始前完成）。",
        "operationId": "stream_hint_v1_tutor__session_id__hint_stream_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TutorHintRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tutor/{session_id}/alt-solutions": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "多解法对比讲解（F-12）",
        "description": "生成 2~3 种解法并做 SymPy 等价性抽检。",
        "operationId": "alt_solutions_v1_tutor__session_id__alt_solutions_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AltSolutionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tutor/{session_id}/analogy": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "生活化类比生成（F-13）",
        "description": "生成贴近学生生活的类比与对应关系说明。",
        "operationId": "analogy_v1_tutor__session_id__analogy_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalogyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tutor/{session_id}/variants": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "举一反三变式题（F-14）",
        "description": "生成 1~2 道同知识点变式题（入库，可作答）。",
        "operationId": "variants_v1_tutor__session_id__variants_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tutor/{session_id}/variants/{variant_id}/answer": {
      "post": {
        "tags": [
          "tutor"
        ],
        "summary": "变式题作答（答错自动回炉，F-14）",
        "description": "答对计入掌握度；答错进入错题本并开启回炉讲解会话。",
        "operationId": "answer_variant_v1_tutor__session_id__variants__variant_id__answer_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "variant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Variant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VariantAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VariantAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AltSolutionOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Steps"
          },
          "scenario": {
            "type": "string",
            "title": "Scenario"
          },
          "answer_verified": {
            "type": "boolean",
            "title": "Answer Verified",
            "description": "SymPy 等价性抽检是否通过"
          }
        },
        "type": "object",
        "required": [
          "title",
          "steps",
          "scenario",
          "answer_verified"
        ],
        "title": "AltSolutionOut",
        "description": "一种解法。"
      },
      "AltSolutionsResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "solutions": {
            "items": {
              "$ref": "#/components/schemas/AltSolutionOut"
            },
            "type": "array",
            "title": "Solutions"
          },
          "checked_count": {
            "type": "integer",
            "title": "Checked Count"
          },
          "verified_count": {
            "type": "integer",
            "title": "Verified Count"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "solutions",
          "checked_count",
          "verified_count"
        ],
        "title": "AltSolutionsResponse",
        "description": "多解法对比。"
      },
      "AnalogyResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "analogy": {
            "type": "string",
            "title": "Analogy"
          },
          "mapping": {
            "type": "string",
            "title": "Mapping"
          },
          "caveat": {
            "type": "string",
            "title": "Caveat"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "analogy",
          "mapping",
          "caveat"
        ],
        "title": "AnalogyResponse",
        "description": "生活化类比。"
      },
...
      "BackToTutorOut": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "message"
        ],
        "title": "BackToTutorOut",
        "description": "答错回炉信息。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TutorHintRequest": {
        "properties": {
          "level": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          }
        },
        "type": "object",
        "title": "TutorHintRequest",
        "description": "请求某层提示（不填则自动进入下一层）。"
      },
      "TutorHintResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "level_name": {
            "type": "string",
            "title": "Level Name"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "next_level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Level"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "level",
          "level_name",
          "content"
        ],
        "title": "TutorHintResponse",
        "description": "提示内容。"
      },
      "TutorQuestionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stem": {
            "type": "string",
            "title": "Stem"
          },
          "qtype": {
            "type": "string",
            "title": "Qtype"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          },
          "knowledge_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Knowledge Points"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stem",
          "qtype",
          "difficulty"
        ],
        "title": "TutorQuestionOut",
        "description": "讲解题目信息（不含答案）。"
      },
      "TutorSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "question": {
            "$ref": "#/components/schemas/TutorQuestionOut"
          },
          "hint_level": {
            "type": "integer",
            "title": "Hint Level"
          },
          "hint_level_name": {
            "type": "string",
            "title": "Hint Level Name"
          },
          "max_hint_level": {
            "type": "integer",
            "title": "Max Hint Level",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "question",
          "hint_level",
          "hint_level_name"
        ],
        "title": "TutorSessionResponse",
        "description": "讲解会话。"
      },
      "TutorSessionStartRequest": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          }
        },
        "type": "object",
        "required": [
          "question_id"
        ],
        "title": "TutorSessionStartRequest",
        "description": "开启讲解会话。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VariantAnswerRequest": {
        "properties": {
          "answer": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "VariantAnswerRequest",
        "description": "变式题作答。"
      },
      "VariantAnswerResponse": {
        "properties": {
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "correct_answer": {
            "type": "string",
            "title": "Correct Answer"
          },
          "back_to_tutor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BackToTutorOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "is_correct",
          "correct_answer"
        ],
        "title": "VariantAnswerResponse",
        "description": "变式题作答结果。"
      },
      "VariantOut": {
        "properties": {
          "question_id": {
            "type": "string",
            "format": "uuid",
            "title": "Question Id"
          },
          "stem": {
            "type": "string",
            "title": "Stem"
          },
          "options": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "stem",
          "difficulty"
        ],
        "title": "VariantOut",
        "description": "变式题（不含答案）。"
      },
      "VariantsResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/VariantOut"
            },
            "type": "array",
            "title": "Variants"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "variants"
        ],
        "title": "VariantsResponse",
        "description": "变式题列表。"
      },
...
    }
  }
}